api/requirements.txt
flask
flask-cors
scikit-learn
xgboost
lightgbm
pandas
numpy
matplotlib
seaborn
joblib
scipy
pillow
python-dotenv

# Azure SDK
azure-storage-blob